type, interfaceの宣言をtypes.tsに集約
